scripts/benchmarking/postgres_migration/requirements.txt
# PostgreSQL Migration Benchmark Dependencies
psycopg2-binary>=2.9.9
pgvector>=0.2.4
chromadb>=0.4.0
numpy>=1.24.0
rank-bm25>=0.2.2
tabulate>=0.9.0
tqdm>=4.66.0
matplotlib>=3.8.0
